Human Radar fix
I have fixed a issue for when the player Cloaks, the player that is cloaked would show up on the radar.

I have added a check if cloaked dont show on radar. This will fix issues of when the a admin needs to go under cover and watch something before it happens